requirements.txt
langchain[all]
langserve[all]
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
seaborn
matplotlib
cohere
mistletoe
mangum
python-dotenv
langchain-groq
wikipedia
langchain-anthropic
langchain-cohere